// front/lib/generated/asset_registry/types/kybStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum KybStatus {
  Pending,
  Verified,
  Rejected,
  Suspended,
}

export type KybStatusArgs = KybStatus;

export function getKybStatusEncoder(): FixedSizeEncoder<KybStatusArgs> {
  return getEnumEncoder(KybStatus);
}

export function getKybStatusDecoder(): FixedSizeDecoder<KybStatus> {
  return getEnumDecoder(KybStatus);
}

export function getKybStatusCodec(): FixedSizeCodec<KybStatusArgs, KybStatus> {
  return combineCodec(getKybStatusEncoder(), getKybStatusDecoder());
}